Refactoring jarvis_convex_hull
I tryied to understand this really intresting algorithm .. but it was a be  confuse for me.
I guess this version is more clear for fresh eyes and respect pep8
Unfortunately, I don't have a good environment to test, but I should work cause I didn't touched the logic. I just changed the naming and the block organization.